prepare to register npm package
