+++ package.json
@@ -5,13 +5,29 @@
   "keywords": [
     "E2D3"
   ],
+  "homepage": "http://e2d3.org/",
+  "bugs": "https://github.com/e2d3/e2d3/issues",
+  "license": "Apache-2.0",
+  "author": "E2D3 Project",
+  "main": "./lib/index.js",
+  "bin": "./bin/e2d3",
+  "scripts": {
+    "postinstall": "bower prune && bower update && gulp build --release",
+    "watch": "gulp run",
+    "start": "gulp run"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/e2d3/e2d3.git"
+  },
   "dependencies": {
     "amd-optimize": "~0.4.3",
     "body-parser": "~1.12.2",
     "bower": "~1.4.1",
     "coffee-script": "~1.9.1",
+    "commander": "^2.8.1",
     "connect-livereload": "~0.5.3",
-    "cors": "^2.6.0",
+    "cors": "~2.6.0",
     "express": "~4.12.3",
     "gulp": "~3.8.11",
     "gulp-coffee": "~2.3.1",
@@ -30,18 +46,12 @@
     "gulp-util": "~3.0.4",
     "main-bower-files": "~2.6.2",
     "merge2": "~0.3.4",
+    "morgan": "~1.5.2",
+    "node-watch": "~0.3.4",
     "rimraf": "~2.3.2",
     "tiny-lr": "~0.1.5"
   },
-  "devDependencies": {},
   "engines": {
     "node": "~0.12.0"
-  },
-  "author": "E2D3 Project",
-  "license": "Apache 2.0",
-  "scripts": {
-    "postinstall": "bower prune && bower update",
-    "watch": "gulp run",
-    "start": "gulp run"
   }
 }
